new tests

+++ heisenberg/pauli.py
@@ -3,10 +3,11 @@ Module for Pauli operators
 """
 import numpy as np
 
-I = np.eye(2)
-X = np.array([[0, 1], [1, 0]])
-Y = np.array([[0, -1j], [1j, 0]])
-Z = np.array([[1, 0], [0, -1]])
+
+I = np.eye(2, dtype=np.complex64)
+X = np.array([[0, 1], [1, 0]], dtype=np.complex64)
+Y = np.array([[0, -1j], [1j, 0]], dtype=np.complex64)
+Z = np.array([[1, 0], [0, -1]], dtype=np.complex64)
 
 pauli_dict = {"I": I, "X": X, "Y": Y, "Z": Z}
